[New] Add command line tool, bls2html, as optional feature.
The command line tool can be used as a useful command line tool to
really do RST-HTML conversion works, as well as a good example of how
to use Birlstone APIs.

It's disabled by default to avoid unncessary dependencies, which is
useful for developers who want to integrate Birlstone as a library in
their own packages.